refactor(CytoscapeRenderer): move renderer logic into separate folder
